Allow pause/resume of hosts
The idea here is to simulate a host *completely* locking up. This could
be due to a hardware fault. When a host locks up, often it doesn't even
know that it did. For example, it may think only 1 second has elapsed
when minutes may have.

In the real world, something like NTP would eventually clue the host in
to the problem.

There is a small test. If you interrogate `Instant::now` on both hosts,
you'll see that the server (which is paused) thinks this test only takes
5ms, while the client thinks it takes over a second.

I'm not sure we actually want to merge this feature yet. I more wrote it
to understand how things were fitting together.

See also: https://questforglory.fandom.com/wiki/Butcher_Shop