Add `Sim::bounce` (#12)
This method drops the `LocalSet` on the host and then restarts the
software (i.e. re-generates the future provided via `Sim::host`). A new
test shows how this can be used.

I had to make `Sim::host` require the factory be `'static`, which kind
of sucks. I don't know a way around that. If anybody has ideas..