Fix error in README.md
In a recent commit, I changed the API for starting a host to require a
future factory. The example in the README broke, but the PR was still
merged.

In this commit, we use Github actions to run `cargo test --verbose` PRs
to main (and adhoc pushes). For now, this just tests on ubuntu as we
have no platform specific code.

In addition, we run the README as a doctest. This slightly harms the
README because the example is no longer copy-paste ready. In particular,
the `#[test] fn simulation() {  }` chrome is now removed so as to work
within the doctest framework. I think that's OK, especially since we
ship real examples in test/.

Finally, the actual bug was fixed by adding a tunnel, down which no
variables slide.