Simulated UDP sockets (#46)
This is a large refactor that includes support for multiple
`SocketAddrs` on each host, IP based links in the network topology,
and more realistic (and correct) message delivery.

Instead of duplicating message delivery logic in the host, it becomes
agnostic to protocol and is encapsulated in `Link`.

Dynamic message passing is removed in favor of more realistic sockets
that deal in bytes.

Networking software is modularized and removed from the host. TCP will
follow the same path in a subsequent change.