Fix host elapsed time across software restarts (#65)
Previously, we simply tracked the elapsed duration after each tick. This makes it impossible to get a host-level view of elapsed time from within the software. Software _could_ use an `Instant` + base `SystemTime` approach, but the `Instant` has issues across host restarts as the tokio runtime is replaced.

This change fixes that by tracking bulk elapsed time on each tick and uses that as a base along with a new `Instant` that is set each time the software runs. With the elapsed base we can accurately give a current elapsed time to running software and survive restarts.