Add network topology introspection (#78)
The sim now supports iterating each host pair's network link, as well
as each message that is currently "in flight" on the network.

Link and message metadata is exposed. We can also now force delivery
of messages.

In combination with the step capabilities that were recently added we
now have the ability to run more complex test scenarios.