Add option to query hosts via regex (#77)
This allows various host interactions via regex matches to reduce the
effort to induce more complex interactions. While this has changed the
explicit API contract in some cases, this should not affect most test
cases as the `ToIpAddrs` trait implements `ToIpAddr` for all `T` that
implement `ToIpAddr`.

This feature is behind a new "regex" feature flag, and the CI is now run
with and without the flag enabled.