Add support for loopback (#118)
This change adds UDP and TCP support for loopback. Packets sent over loopback
do not flow through the topology, instead they are sent directly from the host
back to itself - somewhat modeling how the real world works. The side effect of
this implementation is that network manipulation APIs do not work on loopback
traffic. However, this approach gets incrementally closer towards full loopback
support and unblocks most use cases.

The implementation is a bit of hack. I think where we want to get to is full
support for multiple interfaces. For example, hosts should be allowed to have
both an v4 and v6 ip addr. This will require a bit of a lift due to how hosts
are identified by an IP address.