Fix builder to use provided rng
This commit fixes the turmoil sim builder by removing the `rng`
function. Currently the `build` function ignores the `rng` set on the
builder, and due to the boxing of the value there is no easy way to
repeatedly build a sim with the same `rng`. We would need to add the
ability to clone the boxed rng, which requires some complicated trait
bounds and a different breaking of the interface. This is simpler and
fixes the bug where rng was ignored during building.

This is a breaking change:
- `rng` is removed from the builder, as it was not delivering its
  expected behaviour. Users now need to explicitly specify
  `build_with_rng`.