Add a tracing warning when a thread blocks steps
Add a warning to the sim when a given host or client blocks progress in
a simulation run. This works by spawning a background thread for each
run that periodically checks the steps taken by the simulation. If the
number of steps is the same between checks then the thread adds the
tracing info.