Add established_tcp_stream_count{,_on}
This commit adds a utility method to count the number of established tcp
streams on a given host (or the current host). This can be used to
verify connections have been cleaned up, which is sometimes tricky to
reason about due to a combination of how tcp cleans up and how Drop
works.