Seed tokio with the world rng (#245)
* Introduce `rt::Config` type

In preparation for making tokio's rng seed configurable, introduce a new
`rt::Config` struct that holds configuration for creating tokio
runtimes and pass that around instead of the bare `enable_io` bool.

* Seed tokio runtimes using the world rng

If the `tokio_unstable` flag is set, seed host and client runtimes using
the world rng. This makes the simulation more deterministic. In
particular, the order in which tokio polls `select!` branches is
determined by this seed.