Let network faults affect in-flight messages (#252)
When injecting network faults (partition, hold) into a link, also update
the state of messages already in the `sent` queue, instead of having the
fault only affect messages that are enqueued in the future.

This change enables writing tests in which one observes the contents of
a link to find the point in time in which a given message is sent, to
then prevent that message from getting delivered.