Let TCP peek methods take `&self`
This changes the `TcpStream::peek` and the various other peek methods to
take `&self` instead of `&mut self`, ensuring tokio API compatibility
and making them usable in more contexts.

To achieve this, we wrap the `ReadHalf` internals in `Mutex` and
`AtomicBool`, respectively. Note that we can't use `RefCell`/`Cell` here
because we also require `ReadHalf` to be `Sync`.